+++ src/Picums.Mvc/Configuration/Defaults/BasicDefault.cs
@@ -1,14 +1,18 @@
 ﻿using System.Collections.Generic;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace Picums.Mvc.Configuration.Defaults
 {
     public abstract class BasicDefault : IDefault
     {
+        protected IConfigurationRoot ConfigurationRoot { get; private set; }
+
         public void Apply(StartupConfigurations host, IEnumerable<object> arguments)
         {
+            this.ConfigurationRoot = host.ConfigurationRoot;
             host.ASP.Add((app, env) => this.ConfigureApp(app, env, arguments));
             host.Services.Add(services => this.ConfigureServices(services, arguments));
         }
